Fix for ":return:"/":rtype:" Sphinx markup.
Without the fix, the markup was not properly rendered and was not properly
translated via Sphinx localization resources ("sphinx.po").

The fix is inspired by comparing the code in LuaDomain.py in Tarantool repo
with the code in
https://github.com/redjack/sphinx-lua/blob/develop/src/redjack/sphinx/lua.py

Review (and possibly fixing in some other way) by the author of
LuaDomain.py could be needed.